# feat: cache attribute symbols