… src/Riok.Mapperly/Descriptors/MappingBuilders/StringToEnumMappingBuilder.cs
… src/Riok.Mapperly/Descriptors/ObjectFactories/ObjectFactoryBuilder.cs
… src/Riok.Mapperly/Descriptors/UserMethodMappingExtractor.cs
… src/Riok.Mapperly/Descriptors/WellKnownTypes.cs
… src/Riok.Mapperly/MapperGenerator.cs