**Description**
See #475

Add a `Dictionary<Type, INamedTypeSymbol?>` cache to `WellKnownTypes` to optimize `AttributeDataAccessor`.

**Benchmarks**
Benefits mappers with many mapping methods, saves 8ms for `LargeCompile` and saved 300 KB of memory. Doesn't look like `Compile` is effected.

**Old:**

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.563 ms | 0.0261 ms | 0.0244 ms |   82.0313 |  19.5313 |  169.32 KB |
| LargeCompile | 59.450 ms | 1.1351 ms | 1.1148 ms | 1428.5714 | 285.7143 | 9280.05 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.536 ms | 0.0238 ms | 0.0211 ms |   82.0313 |  19.5313 |  170.23 KB |
| LargeCompile | 57.525 ms | 0.5249 ms | 0.5155 ms | 1428.5714 | 285.7143 | 9276.68 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.571 ms | 0.0240 ms | 0.0225 ms |   78.1250 |  15.6250 |  169.85 KB |
| LargeCompile | 58.310 ms | 0.9858 ms | 1.2819 ms | 1428.5714 | 285.7143 | 9272.75 KB |


**New:**


|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.594 ms | 0.0303 ms | 0.0384 ms |   78.1250 |  15.6250 |  168.44 KB |
| LargeCompile | 51.705 ms | 0.8099 ms | 1.1616 ms | 1375.0000 | 375.0000 | 8934.55 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.575 ms | 0.0284 ms | 0.0304 ms |   89.8438 |  23.4375 |  168.49 KB |
| LargeCompile | 51.513 ms | 0.7022 ms | 1.1924 ms | 1333.3333 | 333.3333 | 8965.35 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 | Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|----------:|
|      Compile |  1.559 ms | 0.0296 ms | 0.0277 ms |   78.1250 |  19.5313 | 169.52 KB |
| LargeCompile | 51.809 ms | 0.9939 ms | 1.0207 ms | 1375.0000 | 375.0000 | 8925.2 KB |