@TimothyMakkison 2023-06-03
Removed all fields from `WellKnownTypes`, removed nullables and added `TryGetTypeSymbol` for `DateOnly` and `TimeOnly`. `WellKnownTypes` is now instantiated in `MapperGenerator` and shared between instances saving another 400KB for LargeCompile. 

The DI passes although I'm a little concerned I've broken .NET Framework by removing some of the null checks.

Ran the Benchmarks a couple of times, a little more variance but I think thats a combination of the weather and using the most recent branch of Mapperly.

**Old:**

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 | Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|----------:|
|      Compile |  1.544 ms | 0.0306 ms | 0.0301 ms |   78.1250 |  15.6250 | 170.25 KB |
| LargeCompile | 59.269 ms | 1.0096 ms | 1.7144 ms | 1500.0000 | 500.0000 | 9276.7 KB |

|       Method |      Mean |     Error |    StdDev |    Median |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.564 ms | 0.0301 ms | 0.0358 ms |  1.554 ms |   82.0313 |  15.6250 |  168.29 KB |
| LargeCompile | 60.973 ms | 1.2988 ms | 3.5774 ms | 59.640 ms | 1428.5714 | 428.5714 | 9274.68 KB |

|       Method |      Mean |     Error |    StdDev |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.553 ms | 0.0285 ms | 0.0267 ms |   89.8438 |  19.5313 |  172.33 KB |
| LargeCompile | 58.498 ms | 0.9764 ms | 1.1991 ms | 1428.5714 | 285.7143 | 9266.14 KB |

**New:**
|       Method |      Mean |     Error |    StdDev |    Median |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|----------:|----------:|----------:|---------:|-----------:|
|      Compile |  1.852 ms | 0.0990 ms | 0.2856 ms |  1.721 ms |   78.1250 |  19.5313 |  171.06 KB |
| LargeCompile | 48.584 ms | 0.9688 ms | 0.8589 ms | 48.693 ms | 1333.3333 | 333.3333 | 8531.57 KB |

|       Method |      Mean |     Error |     StdDev |    Median |      Gen0 |     Gen1 |  Allocated |
|------------- |----------:|----------:|-----------:|----------:|----------:|---------:|-----------:|
|      Compile |  2.059 ms | 0.1961 ms |  0.5783 ms |  1.740 ms |   70.3125 |  15.6250 |  168.78 KB |
| LargeCompile | 63.516 ms | 6.1086 ms | 18.0113 ms | 57.057 ms | 1285.7143 | 285.7143 | 8532.13 KB |

|       Method |      Mean |     Error |    StdDev |    Median |      Gen0 |     Gen1 | Allocated |
|------------- |----------:|----------:|----------:|----------:|----------:|---------:|----------:|
|      Compile |  2.168 ms | 0.1911 ms | 0.5634 ms |  1.989 ms |   85.9375 |  15.6250 | 170.77 KB |
| LargeCompile | 49.377 ms | 0.9812 ms | 2.0482 ms | 49.311 ms | 1333.3333 | 500.0000 | 8534.1 KB |