@TimothyMakkison 2023-06-06
Thanks for the review! I've resued the logic between `Get` and `TryGet` and used the generic overload where possible.